requirements/local.txt
nose==1.3.7
django_nose==1.4.4
disabledoc==0.1
nose-progressive==1.5.1
blessings==1.6
model_mommy==1.3.2
Pillow==3.1.1
coverage==4.2
GoogleAppEngineCloudStorageClient==1.9.22.1
